requirements.txt

torch
torchvision
transformers
sentence-transformers


Pillow

numpy

fastapi
uvicorn
pydantic
python-multipart
tqdm
